+++ app/src/main/java/com/example/gagyeboost/ui/home/HomeFragment.kt
@@ -2,7 +2,9 @@ package com.example.gagyeboost.ui.home
 
 import android.os.Bundle
 import android.view.View
+import android.widget.Toast
 import androidx.navigation.fragment.findNavController
+import androidx.recyclerview.widget.DividerItemDecoration
 import com.example.gagyeboost.R
 import com.example.gagyeboost.databinding.FragmentHomeBinding
 import com.example.gagyeboost.ui.MainViewModel
@@ -13,16 +15,38 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>(R.layout.fragment_home) {
 
     private val viewModel by sharedViewModel<MainViewModel>()
 
+    private val customCalendarAdapter = CustomCalendarAdapter {
+        Toast.makeText(requireContext(), it + "CLICKED", Toast.LENGTH_SHORT).show()
+    }
+
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         binding.viewModel = viewModel
-
+        initView()
         viewModel.getMonthIncome()
         viewModel.getMonthExpense()
         viewModel.setTotalMoney()
-
         binding.fabAdd.setOnClickListener {
             findNavController().navigate(R.id.action_homeFragment_to_addFragment)
+        }
+    }
+
+    private fun initView() {
+
+        with(binding.rvCalendar) {
+            adapter = customCalendarAdapter
+            addItemDecoration(
+                DividerItemDecoration(
+                    requireContext(),
+                    DividerItemDecoration.HORIZONTAL
+                )
+            )
+            addItemDecoration(
+                DividerItemDecoration(
+                    requireContext(),
+                    DividerItemDecoration.VERTICAL
+                )
+            )
         }
     }
 }
